requirements.txt
pyautogui
opencv-python
Pillow
keyboard
pyinstaller
